move functionality to the correct place
